Memindahkan daily motivation ke homepage
