+++ src/scripts/utils/crud-todo.js
@@ -4,18 +4,6 @@ function crudTodo () {
   const TODOS_LOCAL_STORAGE = 'TODOS_LOCAL_STORAGE'
 
   const generateTodoElement = (todoObject) => {
-    // todoCard.innerHTML = `
-    //     <div class="todo-card-left-section d-flex align-items-center w-75">
-    //         <input class="task-checkbox" id="task-${todoObject.id}" type="checkbox">
-    //         <label for="task-${todoObject.id}">
-    //             <span>&#10004;</span>
-    //             <p class="mb-0" id="task-title">${todoObject.task}</p>
-    //         </label>
-    //     </div>
-    // `
-
-    // const checkBoxLabel = document.createElement("span");
-    // checkBoxLabel.innerText ="&#10004;";
     const inputNewTaskElement = document.createElement('input')
     inputNewTaskElement.setAttribute('id', 'new-task')
     inputNewTaskElement.classList.add('input-new-task', 'mb-0')
@@ -63,7 +51,6 @@ function crudTodo () {
       }
 
       labelElement.replaceChild(inputNewTaskElement, todoTitleElement)
-      // todoTitleElement.innerHTML = `<input id="new-task" class="input-new-task mb-0" value="${todoObject.task}">`
       todoCardRightSection.replaceChild(saveTodoBtn, editTodoBtn)
     })
 
@@ -83,13 +70,8 @@ function crudTodo () {
       todos.splice(todoTarget, 1, todoObject)
 
       saveData()
-
-      // inisialisasi element container untuk todo
-      // dan mengosongkan isi element container sebelum looping data terbaru
       const todoListContainer = document.querySelector('.todo-list-container')
       todoListContainer.innerHTML = ''
-
-      // looping semua data todo
       for (const todoItem of todos) {
         const newTodo = generateTodoElement(todoItem)
         todoListContainer.append(newTodo)
@@ -136,13 +118,8 @@ function crudTodo () {
 
       todos.push(todoObject)
       saveData()
-
-      // inisialisasi element container untuk todo
-      // dan mengosongkan isi element container sebelum looping data terbaru
       const todoListContainer = document.querySelector('.todo-list-container')
       todoListContainer.innerHTML = ''
-
-      // looping semua data todo
       for (const todoItem of todos) {
         const newTodo = generateTodoElement(todoItem)
         todoListContainer.append(newTodo)
@@ -151,58 +128,12 @@ function crudTodo () {
       document.getElementById('todo-title-input').value = null
     })
 
-    if (typeof (Storage) !== undefined) {
+    if (checkingStorage) {
       loadDataFromStorage()
     } else {
       alert('Browser kamu tidak mendukung local storage')
     }
   })
-
-  // const generateTodoElement = (todoObject) => {
-  //     const checkBoxLabel = document.createElement("span");
-  //     checkBoxLabel.innerText = `&#10004;`;
-
-  //     const todoTitleElement = document.createElement("p");
-  //     todoTitleElement.classList.add("mb-0");
-  //     todoTitleElement.innerText = todoObject.task;
-
-  //     const labelElement = document.createElement("label");
-  //     labelElement.setAttribute("for", `task-${todoObject.id}`);
-  //     labelElement.append(checkBoxLabel, todoTitleElement);
-
-  //     const checkBoxElement = document.createElement("input");
-  //     checkBoxElement.setAttribute("class", "task-checkbox");
-  //     checkBoxElement.setAttribute("id", `task-${todoObject.id}`);
-  //     checkBoxElement.setAttribute("type", "checkbox");
-
-  //     const todoCardLeftSection = document.createElement("div");
-  //     todoCardLeftSection.setAttribute("class", "todo-card-left-section d-flex align-items-center w-75");
-  //     todoCardLeftSection.append(checkBoxElement, labelElement);
-
-  //     const editTodoIcon = document.createElement("ion-icon");
-  //     editTodoIcon.setAttribute("name", "create-outline");
-
-  //     const deleteTodoIcon = document.createElement("ion-icon");
-  //     deleteTodoIcon.setAttribute("name", "trash-outline");
-
-  //     const editTodoBtn = document.createElement("button");
-  //     editTodoBtn.setAttribute("class", "todo-card-edit-btn me-2 d-flex align-items-center");
-  //     editTodoBtn.append(editTodoIcon);
-
-  //     const deleteTodoBtn = document.createElement("button");
-  //     deleteTodoBtn.setAttribute("class", "todo-card-delete-btn d-flex align-items-center");
-  //     deleteTodoBtn.append(deleteTodoIcon);
-
-  //     const todoCardRightSection = document.createElement("div");
-  //     todoCardRightSection.setAttribute("class", "todo-card-right-section d-flex justify-content-end align-items-center w-25");
-  //     todoCardRightSection.append(editTodoBtn, deleteTodoBtn);
-
-  //     const todoCard = document.createElement("div");
-  //     todoCard.classList.add("todo-card", "p-3", "d-flex", "my-4");
-  //     todoCard.append(todoCardLeftSection, todoCardRightSection);
-
-  //     return todoCard;
-  // }
 
   function findTodoIndex (todoId) {
     for (const index in todos) {
@@ -231,13 +162,8 @@ function crudTodo () {
     }
 
     todos.splice(todoTarget, 1)
-
-    // inisialisasi element container untuk todo
-    // dan mengosongkan isi element container sebelum looping data terbaru
     const todoListContainer = document.querySelector('.todo-list-container')
     todoListContainer.innerHTML = ''
-
-    // looping semua data todo
     for (const todoItem of todos) {
       const newTodo = generateTodoElement(todoItem)
       todoListContainer.append(newTodo)
@@ -246,9 +172,12 @@ function crudTodo () {
     saveData()
   }
 
-  // implementasi web storage
+  function checkingStorage () {
+    return typeof (Storage)
+  }
+
   function saveData () {
-    if (typeof (Storage) !== undefined) {
+    if (checkingStorage) {
       const parsed = JSON.stringify(todos)
       localStorage.setItem(TODOS_LOCAL_STORAGE, parsed)
     } else {
@@ -266,13 +195,9 @@ function crudTodo () {
         todos.push(todo)
       }
     }
-
-    // inisialisasi element container untuk todo
-    // dan mengosongkan isi element container sebelum looping data terbaru
     const todoListContainer = document.querySelector('.todo-list-container')
     todoListContainer.innerHTML = ''
 
-    // looping semua data todo
     for (const todoItem of todos) {
       const newTodo = generateTodoElement(todoItem)
       todoListContainer.append(newTodo)
