+++ src/scripts/views/pages/todopage.js
@@ -1,33 +1,15 @@
-import QuotesDbSource from '../../data/quotesdb-source'
-import { todoPageTemplate, sectionDailyMotivation } from '../templates/template-creator'
+import { todoPageTemplate } from '../templates/template-creator'
 import main from '../../utils/todo'
 import crudTodo from '../../utils/crud-todo'
 
 const TodoPage = {
   async render () {
     return `
-    ${sectionDailyMotivation()}
     ${todoPageTemplate()}
     `
   },
 
   async afterRender () {
-    const quoteText = document.querySelector('.quote')
-    const authorName = document.querySelector('.name')
-    const dailyMotivation = await QuotesDbSource.homePage()
-    quoteText.innerText = dailyMotivation.content
-    authorName.innerText = dailyMotivation.author
-
-    const generateButton = document.querySelectorAll('.generateQuote')
-    generateButton.forEach(button => {
-      button.addEventListener('click', async () => {
-        const nextquoteText = document.querySelector('.quote')
-        const nextauthorName = document.querySelector('.name')
-        const nextMotivation = await QuotesDbSource.homePage()
-        nextquoteText.innerText = nextMotivation.content
-        nextauthorName.innerText = nextMotivation.author
-      })
-    })
     main()
     crudTodo()
   }
