+++ src/styles/style.css
@@ -19,10 +19,6 @@ a{
     text-decoration: none !important;
 }
 
-ion-icon {
-    font-size: 25px;
-}
-
 /* navigation */
 body, header{
     position: sticky;
@@ -178,6 +174,10 @@ footer{
 
 .todo-card-left-section .task-checkbox:checked + label p {
     text-decoration: line-through;
+}
+
+.todo-card-right-section ion-icon{
+    font-size: 25px;
 }
 
 .todo-card-edit-btn {
@@ -649,7 +649,7 @@ button{
 @media screen and (min-width: 750px) {
     app-bar {
         grid-template-columns: 1fr auto;
-        padding: 8px 50px !important;
+        padding: 0 50px !important;
     }
    
     app-bar .app-bar__menu {
