+++ src/templates/index.html
@@ -10,7 +10,7 @@
 <body>
     <header class="bg-secondarypurple"><app-bar></app-bar></header>
     <main id="mainContent"></main>
-    <footer class="text-center text-lg-start text-white bg-altgrey"><app-footer></app-footer></footer>
+    <footer class="text-center text-lg-start text-white bg-secondarypurple mt-5 border-top border-5 border-altpurple"><app-footer></app-footer></footer>
     <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
     <script type="module" src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"></script>
     <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
